Change my master scheme for j2ee unloading
Doesn't make any sense when running from the REPL. So just register a
handler in main() that will stop the system when the container is
unloaded.

Still need a unique name per container. Which, really, needs to come
from the environment. But that's all snarled up with basic system
configuration. Leave it for now, because it makes the REPL code uglier,
and that's the part that is most interesting at the moment.