Have the system almost initializing
This involved a bunch of annoying schema trivia that don't seem worth
the effort at the moment, but I'm fairly sure they'll pay me back
eventually.

One of those is the reason I paused here for a milestone. I'm not sure
about how to specify what system's initialize! actually returns. I
wanted to check the rest of these changes in before I start
experimenting with that.