Just copy/paste a fundamental example
Based on:
https://www.snip2code.com/Snippet/29969/Three-js-demo-with-ClojureScript--Actual

That was posted a little over a year ago, and this version errors out
because (AFAICT) the CanvasRenderer is either obsolete or changed to
something that shouldn't be created directly.

But the basic code seems to be loading and working. So apparently I was
misinterpreting something and being overly pessimistic in my previous checkin