Update namespaces to reflect new file locations
Note that this is still pretty broken: the way I'm loading the basic
system configuration from the environment is still pointing to the old
layout. This seems like a glaring flaw in my current approach.

Then again, I don't think I've really tried to get this working with my
component-dsl. I'm pretty sure this is where I started on that.