Something's drastically wrong
I've spent the past couple of days trying to narrow it down. I don't
have much beyond "Connecting to this will bring your server to its
knees."

Just running the server by itself might have the same effect. It might
possibly be a known issue with Immutant (if I'm lucky and that's the
case, it should be fixed with 2.1...I'm willing to bet the problem's on
my end).

This check-in is almost all comments about trying to isolate the
problem. I don't think there's anything meaningful in here.