This checkin may mark the rope across the gorge
Successfully submitting a request to start auth to the server, and
getting back its totally bogus response.

Don't particularly want most of that response to wind up on the web
browser, but it still feels like a very nice milestone.

Getting an error in the browser that "[object  Object] is not ISeqable,"
but I don't have any context for it (except that it happens while the
event loop is waiting for input from the outside.