Pushing toward a real login exchange
It really needs to be an async event loop or two, interacting with the
client's async event loop(s). This is a couple of baby steps in that
direction.

Actually, this is a back-up of the baby steps I'm trying to take in that
direction. It's cratering my dev environment. I'm about 90% positive the
problem lies outside this code, but I want a backup of my work before I
look at any other approaches.

Yes, this should go in a different branch.

Since no one except me cares about this code, I'm violating my
principles and just checking it in while my system is up and running.