Trying to set up initial world's compiler
This starts by setting up an initial world the same way the rest will be
configured.

The core of that is an initial compiler state.

The plain 'ol empty state is missing things like calls to require and
require macros. So start by declaring a ns and trying to load the
bare-bones basics from the web server.

Have a vague outline for what the server side loader should probably
look like.