Couple more tottering baby steps toward ns loading
Has to happen on the server. Request/responses don't
exactly have to go over web socket, but it seems very silly
to do it over REST.

Except that this is exactly what HTTP is for. So maybe I'm
heading the wrong direction.

Oh well. Even if it's backwards, it's progress