Started converting from schema to spec
Routing is badly broken now, but I've been meaning to give it some love
for quite a while now, and it's not like there's a lot of it yet.

Doesn't compile: I'm getting errors about the time libraries I might
have been planning to use at one point. Not sure where they're being
pulled in, but this is enough for one evening.